requirements.txt
mysqlclient==2.0.3
django-haystack==3.0
django_debug_toolbar==3.2
Django==3.1.3
django_mdeditor==0.1.18
Whoosh==2.7.4
jieba==0.42.1
gunicorn==20.0.4
django-redis==4.12.1
hiredis==1.1.0
